+++ src/ett/ett.h
@@ -1,13 +1,12 @@
 #pragma once
 
 #include "allocator/allocator.h"
-#include "macro.h"
 #include "pool/pool.h"
 #include <box2d/collision.h>
 #include <cglm/types.h>
 
 typedef struct {
-    u64 tag;
+    uint64_t tag;
     vec2 pos;
     mat2 aabb;
 } EttIns;
@@ -23,6 +22,6 @@ extern EttMng ett_mng;
 void ett_mng_init(EttMng *mng, size_t ett_cap);
 void ett_mng_destroy(EttMng *mng);
 
-i32 ett_new(EttMng *mng, vec2 pos, mat2 aabb);
-void ett_remv(EttMng *mng, i32 ett);
-EttIns *ett_get(EttMng *mng, i32 ett);
+size_t ett_new(EttMng *mng, vec2 pos, mat2 aabb);
+void ett_remv(EttMng *mng, size_t ett);
+EttIns *ett_get(EttMng *mng, size_t ett);
